chore(suite-common): simplify suite-sync
chore(suite-common): remove unused suite-sync props

chore(suite-common): remove useSuiteSyncAlerts

chore(suite-common): initSuiteSyncNative

chore(suite-common): useSuiteSync => useToggleSuiteSyncMethods
