+++ suite-native/labeling/src/components/EditableLabelLayout.tsx
@@ -3,10 +3,10 @@ import { useSelector } from 'react-redux';
 
 import type { BottomSheetModalMethods } from '@gorhom/bottom-sheet/lib/typescript/types';
 
-import { selectIsSuiteSyncEnabled } from '@suite-common/suite-sync';
+import { selectIsSuiteSyncEnabled, useToggleSuiteSyncMethods } from '@suite-common/suite-sync';
+import { useAlert } from '@suite-native/alerts';
 import { BottomSheetModal, TextButton, useBottomSheetModal } from '@suite-native/atoms';
 import { Translation } from '@suite-native/intl';
-import { useSuiteSyncAlerts } from '@suite-native/suite-sync';
 
 import { selectIsLabelingEnabled } from '../selectors';
 
@@ -16,11 +16,25 @@ type EditableLabelLayoutParams = {
 };
 
 export const EditableLabelLayout = ({ children, label }: EditableLabelLayoutParams) => {
+    const { showAlert } = useAlert();
     const { bottomSheetRef, openModal, closeModal } = useBottomSheetModal();
-    const isSuiteSyncEnabled = useSelector(selectIsSuiteSyncEnabled);
-    const { showSuiteSyncEnableConfirmationAlert } = useSuiteSyncAlerts();
+    const { enableSuiteSyncIfNeeded } = useToggleSuiteSyncMethods();
 
+    const isSuiteSyncEnabled = useSelector(selectIsSuiteSyncEnabled);
     const isLabelingEnabled = useSelector(selectIsLabelingEnabled);
+
+    const showSuiteSyncEnableConfirmationAlert = (onSuccess: () => void) => {
+        showAlert({
+            title: <Translation id="labeling.enableAlert.title" />,
+            description: <Translation id="labeling.enableAlert.description" />,
+            primaryButtonTitle: <Translation id="labeling.enableAlert.cta" />,
+            onPressPrimaryButton: () => {
+                enableSuiteSyncIfNeeded();
+                onSuccess();
+            },
+            secondaryButtonTitle: <Translation id="generic.buttons.cancel" />,
+        });
+    };
 
     const handleAddLabel = () => {
         if (!isSuiteSyncEnabled) {
