+++ suite-native/module-settings/src/components/ToggleLabelingCard.tsx
@@ -1,17 +1,24 @@
 import { useSelector } from 'react-redux';
 
-import { useSuiteSync } from '@suite-common/suite-sync';
-import { selectSelectedDevice } from '@suite-common/wallet-core';
+import { selectIsSuiteSyncEnabled, useToggleSuiteSyncMethods } from '@suite-common/suite-sync';
+import { useAlert } from '@suite-native/alerts';
 import { TouchableSwitchRow } from '@suite-native/atoms';
 import { Translation } from '@suite-native/intl';
-import { useSuiteSyncAlerts } from '@suite-native/suite-sync';
 
 export const ToggleLabelingCard = () => {
-    const { showSuiteSyncDisableConfirmationAlert } = useSuiteSyncAlerts();
-    const selectedDevice = useSelector(selectSelectedDevice);
-    const { isSuiteSyncEnabled, enableSuiteSyncIfNeeded } = useSuiteSync({
-        device: selectedDevice,
-    });
+    const { showAlert } = useAlert();
+    const isSuiteSyncEnabled = useSelector(selectIsSuiteSyncEnabled);
+    const { enableSuiteSyncIfNeeded, disableSuiteSyncIfNeeded } = useToggleSuiteSyncMethods();
+
+    const showSuiteSyncDisableConfirmationAlert = () => {
+        showAlert({
+            title: <Translation id="labeling.disableAlert.title" />,
+            description: <Translation id="labeling.disableAlert.description" />,
+            primaryButtonTitle: <Translation id="labeling.disableAlert.cta" />,
+            onPressPrimaryButton: disableSuiteSyncIfNeeded,
+            secondaryButtonTitle: <Translation id="generic.buttons.cancel" />,
+        });
+    };
 
     const toggleSuiteSync = () =>
         isSuiteSyncEnabled ? showSuiteSyncDisableConfirmationAlert() : enableSuiteSyncIfNeeded();
