+++ suite-native/module-settings/tsconfig.json
@@ -39,7 +39,6 @@
         { "path": "../navigation" },
         { "path": "../settings" },
         { "path": "../storage" },
-        { "path": "../suite-sync" },
         { "path": "../theme" },
         { "path": "../toasts" },
         { "path": "../trading-residence" },
