+++ suite-native/suite-sync/package.json
@@ -16,9 +16,6 @@
         "@suite-common/suite-sync": "workspace:*",
         "@suite-common/suite-sync-evolu": "workspace:*",
         "@suite-common/suite-sync-storage": "workspace:*",
-        "@suite-common/wallet-core": "workspace:*",
-        "@suite-native/alerts": "workspace:*",
-        "@suite-native/intl": "workspace:*",
         "expo": "~54.0.20",
         "expo-sqlite": "^16.0.8",
         "react-redux": "9.2.0"
